Sort processor methods deterministically
This should provide a more stable input for sorting the steps.

In #49096, we sort the steps based on their dependencies. However, in
some cases, we might have the exact same dependencies and thus the
output order is not exactly stable.

Sorting the input makes sure the output order is always the same.